requirements.txt
tensorflow-cpu==2.11.0
streamlit==1.34.0
gdown==4.7.1
numpy
opencv-python
Pillow
